feat: Add error handling to server and client sockets
Added error handling to server for when a client/user leaves a room.
Server will continue to run and won't break
Added validation to users
for entering correct room name
Moved validation from backend to
frontend, so we don't have to wait for backend to respond. We can know
straight away and save time