feat: Remove unneeded/duplicated code
When trying to change room, I attempt to close/shutdown client socket
from backend. But the problem was the sockets on the client side are
wrapped in a with statement, so when we reach the end of the with
statement, the socket will automatically close.

I believe the issue was since we close the client socket on the backend,
when we reachthe end of the with statement on the cleint, it tries to
send a message to the server (to close the connection), but since the
connection is already closed, it errors out.