feat: Change architecture to use a single port
The architecture is basically the same except the main socket server
will pass the client socket to the child processes (child socket server)
directly instead of letting the client do it. So we still have rooms in
their own process, the difference just being how the child processes
'connect' to the client.

Because we have child processes and not child socket servers, the child
processes just handle room chat (i.e. they just send messages to all the
users in the room). The way they are 'connected' to the clients is from
the main socket server passing the child process the socket where the
client is connected.

So instead of the client being directly connected to child socket
servers, we have child processes that are passed the sockets (clients)
from the main socket server. So we are basically just doing separation
of concerns or modularising our code.